+++ src/mappings/Season2020/TokenTypes/Driver/TeamDrivers.js
@@ -7,6 +7,7 @@ const NoDriver = {
 const KimiRaikkonen = {
     driverId: '7',
     name: 'Kimi Räikkönen',
+    imageName: 'KimiRaikkonen',
     team: Teams.AlfaRomeoRacingORLEN.team,
     description: 'Kimi Räikkönen, otherwise known as "The Iceman", continues 2020 with Alfa Romeo. He has since become the driver with the most Grand Prix starts, surpassing the previous record of 324.'
 };
@@ -79,8 +80,9 @@ const ValtteriBottas = {
 const SergioPerez = {
     driverId: '11',
     name: 'Sergio Pérez',
+    imageName: 'SergioPerez',
     team: Teams.BWTRacingPointF1Team.team,
-    description: 'Sergio Perez began the 2019 season strongly, ending the season in 10th place overall.'
+    description: 'Sergio Pérez began the 2019 season strongly, ending the season in 10th place overall.'
 };
 const LanceStroll = {
     driverId: '18',